requirements (1).txt
fastapi==0.117.1
uvicorn==0.22.0
gradio>=5.46.1
requests==2.31.0
python-dotenv==1.0.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow>=11.3.0
reportlab==4.0.9
python-multipart==0.0.9
